Only delete reserved item with amount 0 if it exists (#320)

+++ website/borrel/models.py
@@ -193,7 +193,8 @@ class ReservationItem(models.Model):
     def save(self, force_insert=False, force_update=False, using=None, update_fields=None):
         """Save reservation item."""
         if self.amount_reserved == 0 and (self.amount_used is None or self.amount_used == 0):
-            self.delete()
+            if self.pk:
+                self.delete()
             return
 
         if not self.product_price_per_unit:
